tests/wpt/web-platform-tests/css/work-in-progress/opera/animations/animation_shorthand_after_pause.html

<!DOCTYPE html>
<html>
<head>
   <title>CSS Animations - animation shorthand after setting pause</title>
   <link rel="author" title="Opera Software ASA" href="http://www.opera.com/">
   <link rel="help" href="https://drafts.csswg.org/css3-animations/#the-animation-shorthand-property-">
   <meta name="flags" content="dom">
   <meta name="assert" content="The animation shorthand does not set
   animation-play-state">
   <script src="/resources/testharness.js"></script>
   <script src="/resources/testharnessreport.js"></script>
   <style type="text/css">
      @keyframes test {
         from {
            background: silver;
         }
         to {
            background: blue;
         }
      }
      .test {
         animation-play-state: paused;
         animation: test 0.2s;
      }
   </style>
   <script type="text/javascript">
      setup({explicit_done:true});
      onload = function() {
         var testName = document.querySelector('title').firstChild.data;
         var div = document.querySelector('div');
         var log = document.querySelectorAll('p')[0].firstChild;
         div.className = 'test';
         var style = getComputedStyle(div, null);
         var playState = style.getPropertyValue('animation-play-state');
         var passed = (playState == 'paused');
         test(function(){assert_true(passed);}, testName); done();
         log.data = (passed ? 'PASS' : 'FAIL');
      }
   </script>
   <style>iframe{display:none}</style>
</head>
<body>
   <p>Testing... (note: JavaScript required)</p>
   <div></div>
</body>
</html>
